requirements.txt
aiohttp==3.9.3
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.3.0
attrs==23.2.0
beautifulsoup4==4.12.3
certifi==2024.2.2
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
cryptography==42.0.5
cssselect==1.2.0
dataclasses-json==0.6.4
datasets==2.18.0
dill==0.3.8
distro==1.9.0
docopt==0.6.2
faiss-cpu==1.8.0
filelock==3.13.1
frozenlist==1.4.1
fsspec==2024.2.0
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.4
httpx==0.27.0
huggingface-hub==0.21.4
idna==3.6
Jinja2==3.1.3
jsonpatch==1.33
jsonpointer==2.4
langchain==0.1.12
langchain-community==0.0.28
langchain-core==0.1.32
langchain-openai==0.0.8
langchain-text-splitters==0.0.1
langsmith==0.1.31
lxml==5.1.0
MarkupSafe==2.1.5
marshmallow==3.21.1
mpmath==1.3.0
multidict==6.0.5
multiprocess==0.70.16
mypy-extensions==1.0.0
networkx==3.2.1
numpy==1.26.4
openai==1.14.2
orjson==3.9.15
outcome==1.3.0.post0
packaging==23.2
pandas==2.2.1
pdfminer==20191125
pdfminer.six==20231228
pdfquery==0.4.3
pyarrow==15.0.2
pyarrow-hotfix==0.6
pycparser==2.21
pycryptodome==3.20.0
pydantic==2.6.4
pydantic_core==2.16.3
PyMuPDF==1.23.26
PyMuPDFb==1.23.22
pypdf==4.1.0
pyquery==2.0.0
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
regex==2023.12.25
requests==2.31.0
roman==4.1
safetensors==0.4.2
selenium==4.18.1
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.5
SQLAlchemy==2.0.28
sympy==1.12
tenacity==8.2.3
tiktoken==0.6.0
tokenizers==0.15.2
torch==2.2.1
tqdm==4.66.2
transformers==4.38.2
trio==0.25.0
trio-websocket==0.11.1
txtai==7.0.0
txtinstruct==0.1.0
typing-inspect==0.9.0
typing_extensions==4.10.0
tzdata==2024.1
urllib3==2.2.1
wsproto==1.2.0
xxhash==3.4.1
yarl==1.9.4
